refactor(react): extract non-mixin methods
