+++ packages/react/mixins/mixin.server.js
@@ -1,12 +1,9 @@
 'use strict';
 
-const { extname } = require('path');
 const { parse } = require('url');
 
 const { createElement } = require('react');
-const { renderToString } = require('react-dom/server');
 const { default: StaticRouter } = require('react-router-dom/es/StaticRouter');
-const { Helmet } = require('react-helmet');
 
 const {
   override,
@@ -15,6 +12,7 @@ const {
 
 const { Mixin } = require('@untool/core');
 
+const render = require('../lib/render');
 const template = require('../lib/template');
 
 class ReactMixin extends Mixin {
@@ -25,42 +23,6 @@ class ReactMixin extends Mixin {
     const modules = (this.modules = []);
     const context = (this.context = { modules });
     this.options = { ...(options && options.router), basename, context };
-  }
-  determineAssets() {
-    const { entryFiles, vendorFiles, moduleFileMap } = this.stats;
-    const moduleFiles = this.modules.reduce(
-      (result, module) => [...result, ...moduleFileMap[module]],
-      []
-    );
-    return [
-      ...vendorFiles.sort((a, b) => b.localeCompare(a)),
-      ...moduleFiles.sort((a, b) => b.localeCompare(a)),
-      ...entryFiles.sort((a, b) => b.localeCompare(a)),
-    ]
-      .filter(
-        (asset, index, self) =>
-          self.indexOf(asset) === index &&
-          /\.(css|js)$/.test(asset) &&
-          !/\.hot-update\./.test(asset)
-      )
-      .reduce(
-        (result, asset) => {
-          const extension = extname(asset).substring(1);
-          result[extension].push(asset);
-          return result;
-        },
-        { css: [], js: [] }
-      );
-  }
-  getInitialTemplateData({ helmet, ...data }) {
-    const { name: mountpoint, _env } = this.config;
-    const assets = this.determineAssets();
-    const fragments = Object.entries(helmet).reduce(
-      (result, [key, value]) => ({ ...result, [key]: value.toString() }),
-      { headPrefix: '', headSuffix: '' }
-    );
-    const globals = { _env };
-    return { ...data, mountpoint, assets, fragments, globals };
   }
   bootstrap(req, res) {
     const { pathname, search } = parse(req.url);
@@ -75,14 +37,11 @@ class ReactMixin extends Mixin {
       .then(() => this.bootstrap(req, res))
       .then(() => this.enhanceElement(this.element))
       .then((element) =>
-        this.fetchData({}, element).then((data) => ({ element, data }))
+        this.fetchData({}, element).then((data) =>
+          render(element, data, this.config, this.modules, this.stats)
+        )
       )
-      .then(({ element, data: fetchedData }) => {
-        const initialData = this.getInitialTemplateData({
-          markup: renderToString(element),
-          helmet: Helmet.renderStatic(),
-          fetchedData,
-        });
+      .then((initialData) => {
         if (this.context.miss) {
           next();
         } else if (this.context.url) {
